lectures/1/requirements.txt
fastapi
uvicorn
requests
python-multipart
websockets
openai
python-dotenv